Modified ChangeUserDataTest
