+++ src/test/java/UserAuthorizationTest.java
@@ -1,3 +1,4 @@
+import data.Credentials;
 import io.qameta.allure.Step;
 import io.qameta.allure.junit4.DisplayName;
 import io.restassured.response.Response;
@@ -5,6 +6,9 @@ import org.apache.http.HttpStatus;
 import org.junit.Before;
 import org.junit.Test;
 import pogo.AuthResponse;
+import user.UserClient;
+import utils.Specification;
+
 import static io.restassured.RestAssured.given;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.is;
