+++ src/test/java/user/UserGenerator.java
@@ -1,8 +1,10 @@
+package user;
+
 import com.github.javafaker.Faker;
 import java.util.Locale;
 
 
- class UserGenerator {
+ public class UserGenerator {
     Faker faker = new Faker(new Locale("en-RU"));
     public String fakeUserEmail(){
         return faker.internet().emailAddress();}
