+++ src/test/java/utils/Specification.java
@@ -1,3 +1,5 @@
+package utils;
+
 import io.restassured.RestAssured;
 import io.restassured.builder.RequestSpecBuilder;
 import io.restassured.builder.ResponseSpecBuilder;
